Update prepare-item-response for plugin endpoint
Fixes #20. Currently the plugin endpoint's `prepare_item_for_response`
method is not tied to the schema.  It should be for future bc
compatability along with other benefits of this approach.  Currently
many other core endpoints are following this pattern.

This PR also fixes the plugin checks for data to reflect that data being
prepared for the response indeed matches data from the actual plugin
files.